Fixed c_long->int issue in std.socket – the code was probably written back when D on x86_64 was a faint dream.
Actually, the timeval struct could be removed altogether as it isn't used anywhere in std.socket (I didn't realize that I wasn't using the C standard library one), but this would be still a breaking API change, so I refrained from it.

+++ std/socket.d
@@ -40,6 +40,7 @@ module std.socket;
 import core.stdc.stdint, std.string, std.c.string, std.c.stdlib, std.conv,
     std.traits;
 
+import core.stdc.config;
 import core.time : dur, Duration;
 import std.algorithm : max;
 import std.exception : assumeUnique, enforce;
@@ -947,8 +948,8 @@ enum SocketFlags: int
 extern(C) struct timeval
 {
         // D interface
-        int seconds;            /// Number of seconds.
-        int microseconds;       /// Number of additional microseconds.
+        c_long seconds;            /// Number of seconds.
+        c_long microseconds;       /// Number of additional microseconds.
 
         // C interface
         deprecated
